pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "layout-clustering-final"
version = "1.0.0"
description = "Layout clustering and recommendation system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.9.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "loguru>=0.6.0",
    "pytest>=7.0.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "umap-learn>=0.5.4",
    "numba>=0.57.0",
    "llvmlite>=0.40.0",
    "openai>=1.100.2",
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
]

[tool.hatch.build.targets.wheel]
packages = [
    "analysis",
    "clustering",
    "core",
    "embeddings",
    "features",
    "recommendation",
    "utils"
]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
# Enable less strict rules for development
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]

# Ignore some rules that are too strict for development
ignore = [
    "E501",  # Line too long (handled by formatter)
    "UP035", # typing.Dict -> dict (less strict)
    "UP006", # Use list instead of List (less strict)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
# Development-friendly MyPy settings
python_version = "3.8"
warn_return_any = false
warn_unused_configs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_calls = false
disallow_any_generics = false
warn_no_return = false
ignore_missing_imports = true
warn_unused_ignores = false
disable_error_code = [
    "import-untyped",
    "assignment",
    "operator",
    "index",
    "arg-type",
    "return-value",
    "has-type",
    "misc"
]

# Exclude test files and experimental code
exclude = [
    "tests/",
    "conftest.py",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.3.0",
    "ruff>=0.12.9",
    "types-pyyaml>=6.0.12.20250809",
]
